[CSPM] refactor the check cli definition
