+++ pkg/compliance/cli/resolver_security_agent.go
@@ -3,10 +3,9 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-//go:build !windows && !kubeapiserver
+//go:build unix && !kubeapiserver
 
-// Package check holds check related files
-package check
+package cli
 
 import (
 	"github.com/DataDog/datadog-agent/pkg/compliance"
